feat: viewer access to menu planning
