+++ src/proxy.ts
@@ -1,8 +1,16 @@
 import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
 
-const isPublicRoute = createRouteMatcher(["/sign-in(.*)"]);
+const isPublicRoute = createRouteMatcher(["/sign-in(.*)", "/menuPlanning(.*)"]);
 
 export default clerkMiddleware(async (auth, request) => {
+  const { method, url } = request;
+  const pathname = new URL(url).pathname;
+
+  // Allow GET requests to /api/calendar and subroutes without auth
+  if (method === "GET" && /^\/api\/calendar(\/.*)?$/.test(pathname)) {
+    return;
+  }
+
   if (!isPublicRoute(request)) {
     await auth.protect();
   }
